+++ sdk/rollup.config.js
@@ -16,6 +16,7 @@ export default [
             'index.native': 'src/index.native.ts',
             'physics/index': 'src/physics/index.ts',
             'physics3d/index': 'src/physics3d/index.ts',
+            'douyin/index': 'src/platform/douyin/index.ts',
             'spine/index': 'src/spine/index.ts',
             'dragonbones/index': 'src/dragonbones/index.ts',
             'wasm': 'src/wasm.ts',